- fix negative PTS

+++ Daniel2.SimplePlayerGL/ReadFileDN2.cpp
@@ -204,7 +204,7 @@ int ReadFileDN2::ReadFrame(size_t frame, C_Buffer & buffer, size_t & size, size_
 
 		DWORD ret_size = 0;
 
-		if (!SUCCEEDED(m_fileMvx->UnwrapFrame(buffer.GetPtr(), size, 0, &ret_size)))
+		if (!SUCCEEDED(m_fileMvx->UnwrapFrame(buffer.GetPtr(), static_cast<DWORD>(size), 0, &ret_size)))
 			return -1;
 
 		size = ret_size;
